// types.ts
export interface NutritionData {
  servingSize?: string;
  calories?: string;
  sugarContent?: string;
  sodiumContent?: string;
  fatContent?: string;
  saturatedFatContent?: string;
  unsaturatedFatContent?: string;
  transFatContent?: string;
  carbohydrateContent?: string;
  fiberContent?: string;
  proteinContent?: string;
  cholesterolContent?: string;
}

export interface RecipeData {
  name: string;
  description: string;
  prep_time: string;
  cook_time: string;
  total_time: string;
  yield: string;
  keywords: string[];
  cuisine: string;
  category: string;
  ingredients: string[];
  instructions: string[];
  notes?: string;
  image?: string; // base64 string or URL
  image_alt?: string;
  image_title?: string;
  image_description?: string;
  method?: string;
  diet?: string;
  video_url?: string;
  nutrition?: NutritionData;
  aggregateRating?: {
    ratingValue: string;
    ratingCount: number;
  };
}

export interface GeneratedPost {
  post_title: string;
  post_content: string;
  recipe_data: RecipeData;
  focus_keyword?: string;
  meta_description?: string;
  slug?: string;
  faqSchema?: any;
}

export interface WordPressSite {
  id: string;
  name: string;
  url: string;
  siteToken: string;
}

export interface RecipeDetails {
  has_card: boolean;
  has_ingredients: boolean;

  has_instructions: boolean;
  recipe_id?: number | null;
}

export interface WordPressPost {
  id: number;
  title: string;
  link: string;
  recipe_details: RecipeDetails;
  post_status: string;
  has_recipe?: boolean;
  featured_image_url?: string | null;
  // New fields for detailed analysis
  word_count: number;
  headline_counts: { [key: string]: number };
  image_count: number;
  focus_keyword: string | null;
  language: string;
  has_meta_description: boolean;
  has_recipe_schema: boolean;
  has_faq_schema: boolean;
}